+++ Space Defence/SFML Project/Health.h
@@ -2,12 +2,12 @@
 class Health
 {
 private:
-	short healthPoints;
+	int healthPoints;
 public:
-	Health(const short healthPoints = 3);
+	Health(const int healthPoints = 3);
 	Health(const Health &originalHealth);
 	Health& operator=(const Health &originalHealth);
-	short getHealthPoints() const;
+	int getHealthPoints() const;
 	void reduceHealthPoints();
 	bool isAlive() const;
 	~Health();
